Split out MicroserviceGlobal into separate files
