+++ app/uk/gov/hmrc/timetopay/arrangement/config/GuiceModule.scala
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2017 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.timetopay.arrangement.config
+
+import javax.inject.{Inject, Provider}
+import com.google.inject.{AbstractModule, Singleton}
+import uk.gov.hmrc.play.config._
+
+class GuiceModule extends AbstractModule with ServicesConfig {
+  override def configure: Unit = {
+    bind(classOf[reactivemongo.api.DB]).toProvider(classOf[MongoDbProvider])
+    ()
+  }
+}
